# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MTG Deckbuilding API",
    "description": "Commander utility endpoints including deck validation and EDHRec tooling.",
    "version": "1.1.0"
  },
  "paths": {
...
    "/api/v1/cedh/search": {
      "get": {
        "tags": [
          "cedh"
        ],
        "summary": "Search cEDH Database",
        "description": "Search the competitive EDH decklist database with various filters.\n    \n    **Filter Options:**\n    - `commander`: Search by commander name (partial match, case-insensitive)\n    - `colors`: Exact color identity match (e.g., \"ub\" for Dimir, \"wubrg\" for 5-color)\n    - `section`: Filter by competitive level (COMPETITIVE, DEPRECATED, BREW)\n    - `primer_only`: Only return decks with primers available\n    - `limit`: Maximum number of results to return (default: 20)\n    \n    **Examples:**\n    - `/api/v1/cedh/search?commander=tymna` - Find all Tymna decks\n    - `/api/v1/cedh/search?colors=ub&section=COMPETITIVE` - Competitive Dimir decks\n    - `/api/v1/cedh/search?primer_only=true&limit=10` - Top 10 decks with primers",
        "operationId": "search_cedh_decks_api_v1_cedh_search_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "commander",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Commander name to search for",
              "title": "Commander"
            },
            "description": "Commander name to search for"
          },
          {
            "name": "colors",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Color identity (e.g., 'ub', 'wubrg')",
              "title": "Colors"
            },
            "description": "Color identity (e.g., 'ub', 'wubrg')"
          },
          {
            "name": "section",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Section filter: COMPETITIVE, DEPRECATED, or BREW",
              "title": "Section"
            },
            "description": "Section filter: COMPETITIVE, DEPRECATED, or BREW"
          },
          {
            "name": "primer_only",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Only return decks with primers",
              "default": false,
              "title": "Primer Only"
            },
            "description": "Only return decks with primers"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Maximum number of results",
              "default": 20,
              "title": "Limit"
            },
            "description": "Maximum number of results"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/cedh/commanders": {
      "get": {
        "tags": [
          "cedh"
        ],
        "summary": "List All Commanders",
        "description": "Get a list of all commanders in the cEDH database with deck counts.",
        "operationId": "list_commanders_api_v1_cedh_commanders_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/cedh/stats": {
      "get": {
        "tags": [
          "cedh"
        ],
        "summary": "Database Statistics",
        "description": "Get statistics about the cEDH database including color distribution, section breakdown, and more.",
        "operationId": "get_database_stats_api_v1_cedh_stats_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/cedh/info": {
      "get": {
        "tags": [
          "cedh"
        ],
        "summary": "API Information",
        "description": "Get information about the cEDH database API endpoints and usage.",
        "operationId": "get_api_info_api_v1_cedh_info_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}